+++ src/Laravel/Filter/WhereHas.php
@@ -12,9 +12,11 @@ use function Tobyz\JsonApiServer\apply_filters;
 
 class WhereHas extends Filter
 {
-    protected const QUERY_BUILDER_METHOD = 'whereHas';
+    use SupportsOperators;
 
     public Relationship|string|null $field = null;
+
+    public const SUPPORTED_OPERATORS = ['eq', 'in', 'ne', 'notin', 'null', 'notnull'];
 
     public static function make(string $name): static
     {
@@ -38,7 +40,6 @@ class WhereHas extends Filter
             );
         }
 
-        $value = (array) $value;
         $field =
             $this->field instanceof Relationship
                 ? $this->field
@@ -55,23 +56,57 @@ class WhereHas extends Filter
 
         $relatedCollection = $context->api->getCollection($field->collections[0]);
 
-        $query->{static::QUERY_BUILDER_METHOD}($field->property ?: $field->name, function (
-            $query,
-        ) use ($value, $relatedCollection, $context) {
-            if ($relatedCollection instanceof EloquentResource) {
-                $relatedCollection->scope($query, $context);
-            }
+        $value = $this->resolveOperators($value);
 
-            if (array_is_list($value)) {
-                $query->whereKey(array_merge(...array_map(fn($v) => explode(',', $v), $value)));
-            } else {
+        foreach ($value as $operator => $v) {
+            $method = match ($operator) {
+                'ne', 'notin' => 'whereDoesntHave',
+                'null', 'notnull' => ($operator === 'null' xor
+                !filter_var($v, FILTER_VALIDATE_BOOLEAN))
+                    ? 'whereDoesntHave'
+                    : 'whereHas',
+                default => 'whereHas',
+            };
+
+            $query->{$method}($field->property ?: $field->name, function ($query) use (
+                $operator,
+                $v,
+                $relatedCollection,
+                $context,
+            ) {
+                if ($relatedCollection instanceof EloquentResource) {
+                    $relatedCollection->scope($query, $context);
+                }
+
+                if (in_array($operator, ['null', 'notnull'])) {
+                    return;
+                }
+
+                if ($ids = $this->extractIds($v)) {
+                    $query->whereKey($ids);
+                    return;
+                }
+
                 apply_filters(
                     $query,
-                    $value,
+                    (array) $v,
                     $relatedCollection,
                     $context->withCollection($relatedCollection),
                 );
-            }
-        });
+            });
+        }
+    }
+
+    private function extractIds(array|string $value): ?array
+    {
+        if (is_string($value)) {
+            return explode(',', $value);
+        }
+
+        if (array_is_list($value)) {
+            return $value;
+        }
+
+        return null;
     }
 }